feat: allow local admin to set custom bwrap args
This means that, for example, Trivalent could be configured to be unable
to read certain directories.

Also edit a couple comments for accuracy: the GTK icon issue was
determined to be unrelated to Glycin, and the conf files in
`/etc/trivalent/trivalent.conf.d` are more precisely described as
"configured by the local administrator" than "user-configured".